requirements.txt
mysql-connector-python>=8.0.0
customtkinter>=5.0.0

